mcp-client-ui/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
boto3==1.40.11
botocore==1.40.11
certifi==2025.8.3
cffi==1.17.1
chainlit==2.6.9
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cyclopts==3.22.5
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.11.3
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.12.1
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
Lazify==0.4.0
lazy-object-proxy==1.11.0
litellm==1.75.8
literalai==0.1.201
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.2
llama-index-cli==0.5.0
llama-index-core==0.13.2
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.2
llama-index-instrumentation==0.4.0
llama-index-llms-litellm==0.6.0
llama-index-llms-openai==0.5.4
llama-index-readers-file==0.5.2
llama-index-readers-llama-parse==0.5.0
llama-index-tools-mcp==0.4.0
llama-index-workflows==1.3.0
llama-parse==0.6.54
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.7.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.45.6
opentelemetry-instrumentation-anthropic==0.45.6
opentelemetry-instrumentation-bedrock==0.45.6
opentelemetry-instrumentation-chromadb==0.45.6
opentelemetry-instrumentation-cohere==0.45.6
opentelemetry-instrumentation-crewai==0.45.6
opentelemetry-instrumentation-google-generativeai==0.45.6
opentelemetry-instrumentation-groq==0.45.6
opentelemetry-instrumentation-haystack==0.45.6
opentelemetry-instrumentation-lancedb==0.45.6
opentelemetry-instrumentation-langchain==0.45.6
opentelemetry-instrumentation-llamaindex==0.45.6
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.45.6
opentelemetry-instrumentation-mcp==0.45.6
opentelemetry-instrumentation-milvus==0.45.6
opentelemetry-instrumentation-mistralai==0.45.6
opentelemetry-instrumentation-ollama==0.45.6
opentelemetry-instrumentation-openai==0.45.6
opentelemetry-instrumentation-openai-agents==0.45.6
opentelemetry-instrumentation-pinecone==0.45.6
opentelemetry-instrumentation-qdrant==0.45.6
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.45.6
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.45.6
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.45.6
opentelemetry-instrumentation-transformers==0.45.6
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.45.6
opentelemetry-instrumentation-watsonx==0.45.6
opentelemetry-instrumentation-weaviate==0.45.6
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.12
opentelemetry-util-http==0.55b1
packaging==25.0
pandas==2.2.3
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
platformdirs==4.3.8
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
s3transfer==0.13.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.2
striprtf==0.0.26
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.45.6
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.20.0
Werkzeug==3.1.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
